Moving errors in same file
