--- lib/domain/models/NotFound.js
@@ -1,8 +1,0 @@
-class NotFound extends Error {
-  constructor (message = 'Not found') {
-    super(message)
-    this.statusCode = 404
-  }
-}
-
-module.exports = NotFound
