+++ lib/domain/models/errors.js
@@ -5,4 +5,11 @@ class KnownUserError extends Error {
   }
 }
 
-module.exports = KnownUserError
+class NotFound extends Error {
+  constructor (message = 'Not found') {
+    super(message)
+    this.statusCode = 404
+  }
+}
+
+module.exports = { KnownUserError, NotFound }
