+++ lib/domain/services/user-service.js
@@ -1,14 +1,14 @@
-const KnowUserError = require('../models/errors')
+const { KnownUserError } = require('../models/errors')
 
 module.exports = {
   async canUserRegisterByEmailOrPSeudo ({ email, pseudo }, userRepository) {
     const isUserknownByEmail = await userRepository.isUserknownByEmail(email)
     const isUserknownByPseudo = await userRepository.isUserknownByPseudo(pseudo)
     if (isUserknownByEmail) {
-      throw new KnowUserError('email already registered')
+      throw new KnownUserError('email already registered')
     }
     if (isUserknownByPseudo) {
-      throw new KnowUserError('pseudo already registered')
+      throw new KnownUserError('pseudo already registered')
     }
     return true
   }
