+++ lib/domain/useCases/search-users.js
@@ -1,4 +1,4 @@
-const NotFound = require('../models/NotFound')
+const { NotFound } = require('../models/errors')
 
 module.exports = async function searchUsers ({
   search,
