+++ lib/infrasctucture/repositories/user-repository.js
@@ -1,6 +1,6 @@
 const User = require('../../db/models/User')
 const ApiResponse = require('../../domain/models/ApiResponse')
-const NotFound = require('../../domain/models/NotFound')
+const { NotFound } = require('../../domain/models/errors')
 
 module.exports = {
   async isUserknownByPseudo (pseudo) {
