+++ lib/test/api/integration/infrasctructure/repositories/user-repository.test.js
@@ -4,7 +4,7 @@ const NewUser = require('../../../../../domain/models/NewUser')
 const userRepository = require('../../../../../infrasctucture/repositories/user-repository')
 const { expect } = require('chai')
 const mongoose = require('mongoose')
-const NotFound = require('../../../../../domain/models/NotFound')
+const { NotFound } = require('../../../../../domain/models/errors')
 const catchErr = require('../../../../test-helper')
 
 after(async () => {
