+++ lib/test/api/unit/domain/services/user-service.test.js
@@ -9,7 +9,7 @@ chai.use(require('sinon-chai'))
 // Other
 const userRepository = require('../../../../../infrasctucture/repositories/user-repository')
 const service = require('../../../../../domain/services/user-service')
-const KnownUserError = require('../../../../../domain/models/errors')
+const { KnownUserError } = require('../../../../../domain/models/errors')
 const catchErr = require('../../../../test-helper')
 
 describe('Unit | Api | Services | userService', () => {
