+++ lib/test/api/unit/domain/useCases/register-new-user.test.js
@@ -10,7 +10,7 @@ chai.use(require('sinon-chai'))
 const registerNewUser = require('../../../../../domain/useCases/register-new-user')
 const userRepository = require('../../../../../infrasctucture/repositories/user-repository')
 const userService = require('../../../../../domain/services/user-service')
-const KnownUserError = require('../../../../../domain/models/errors')
+const { KnownUserError } = require('../../../../../domain/models/errors')
 
 describe('Unit | Api | UseCase | register-new-user', () => {
   describe('#registerNewUser', () => {
